udev: Updated Dragonrise 0079:0011
There are three separate configuration files for Dragonrise 0079:0011.
Two of them used to have identical button mappings but one of them used
the VID:PID string "Dragonrise Inc. Gamepad" instead of the proper
iProduct string "USB Gamepad " as device_name.

This was corrected and button layout was changed to a more common layout
in that variant of the configuration file.