Fix DragonRise N64 controller mapping
Remapped based on
(1) the auto-config obtained from Lakka.tv on Raspberry Pi 3, and
(2) manually fixing the inputs with the help of https://html5gamepad.com/.

The old config incorrectly mapped "input_a_btn" and "input_b_btn" to the physical buttons "A" and "B".  The above config operates correctly in-game on my rig.