Fix up a few more bad button bindings
I figured out how to edit controller autoconfig without root, so here I think are most of the changes needed to make this functional. Unless specified, I tested the button changes in a game.

- Fix typo to make X and Y buttons work correctly
- There's no select button, so bind it to the back button like on the Nvidia Shield controller
- Change R2 trigger (done through UI's binding system, untested because I don't have a game loaded that uses it yet, but it also matches the Shield controller's binding)
- add and rename some labels for clarity and correctness

There's one remaining problem that the provided start button code is the default KEYCODE_BUTTON_START, but this gamepad has no start button. The middle button should have a keycode that works, but it's being flaky for me, so I've left that alone for now.