Add dinput profiles for 8Bitdo N30 Pro USB/BT.
When N30 Pro is connected/paired in dinput mode (B+power) but RA drivers are set to "xinput", it doesn't read the existing autoconfig from the "xinput" folder for some reason, in BT it uses M30 profile with incorrect inputs, in USB it says "not configured".

So I just copied the profiles from the "xinput" folder to the "dinput" one (and changed the `input_driver` line of course), and now it works fine.