Update sdl2 config for Nintendo Switch Pro Controller
With the current config D-Pad keys do not work. Mappings in this update
are mirroring mappings in the updated udev config for this controller.
These mappings are tested with the new hid-nintendo driver, which
supports both wireless and USB connections of the controller. This
driver was supposed to be included in the kernel from 5.10, but it was
not. The easiest way to get it is to install dkms-hid-nintendo
(https://github.com/nicman23/dkms-hid-nintendo).