Create BrawlerGen USB Genesis Mini (Switch mode).cfg
New Retro Fighters controller. When connected to a USB 2.0 or lower port, it will automatically go into Switch mode. The button labels have been written with respect to what is physically written on the controller despite it using the Switch Pro Controller inputs, making this transparent to the user. Supports the macros as well as the virtual home button (Start+A).