Swap buttons used as A and B
The existing config used the back button on this remote as "A" and the center button as "B". In the default UI, this meant that you need to hit the back button to select an option and the center button to back out. This seemed backwards, so I swapped them.

My guess is that whoever uploaded this configuration was using Nintendo-style controls in their RetroArch UI. Since that wasn't the default in my setup, I don't think it should be the default here. I understand this is a weird case where the remote doesn't resemble a real controller, though, and maybe I'm wrong if the RetroArch default orientation ("Menu Swap OK and Cancel Buttons") changes in cases I'm not aware of.

I could probably go on about a deeper code change that might be a more correct fix that works in all cases, but I don't know that there's interest or that it'd be worth it, so I just proposed this.