Create Wireless Gamepad.cfg
Retro Fighters Brawler64 Wireless Gamepad
UPC: 850046080067


The physical B button shares 2 functions: in game, it serves as the n64 "B" button. in the retroarch menus it serves as a back button, ("A" according to port 1 control menu). 
This is the best way i could figure out how to map this gamepad for nintendo 64 games.
Only problem is that the "Y" button is also serves for deleting hotkeys in the menu.
work around for this is to temporarily set "y" to a C up, down, left, or right button. remove the hotkey errors and then resetting to default controls afterwards in port 1 control.